% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boundary_plot.R
\name{boundary_plot}
\alias{boundary_plot}
\title{Convience function for creating boundary plots}
\usage{
boundary_plot(model, x, y, z, ...)
}
\arguments{
\item{model}{A caret classification model}

\item{x}{Data}

\item{y}{Data}

\item{z}{Observed values}

\item{...}{Additional arguments passed to the contour function}
}
\description{
Convience function for creating boundary plots
}

