% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nclRpredictive-package.R
\docType{package}
\name{nclRpredictive-package}
\alias{nclRpredictive}
\alias{nclRpredictive-package}
\title{R courses at Newcastle University}
\description{
Functions and datasets used for the predictive analytics course at Newcatle University
}
\keyword{package}

