% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot3d.R
\name{plot3d}
\alias{plot3d}
\title{This is a function for assisting with 3d plots for the predictive analytics course}
\usage{
plot3d(model, xvar, yvar, zvar, points = FALSE, ...)
}
\arguments{
\item{model}{A model object as returned by the caret::train function}

\item{xvar}{The variable to be plotted as the x axis}

\item{yvar}{The variable to be plotted as the y axis}

\item{zvar}{The response variable being predicted}

\item{points}{A logical argument if TRUE the resultant plot will also
show the observed points around the fitted surface with red points
highlighting those for which the model under estimates and blue points
for those points which are below the surface.
Note this can get messy when there are large numbers of points.}

\item{...}{Other parameters to be passed to the persp function}
}
\description{
This is a function for assisting with 3d plots for the predictive analytics course
}
\examples{
data(advertising)
m = caret::train(Sales~TV + Radio, data = advertising, method = "lm")
op = par(mar = c(1,1,1,1))
plot3d(m,advertising$TV, advertising$Radio, advertising$Sales, points = TRUE)
par(op)
}

